% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccamlr_mpa}
\alias{ccamlr_mpa}
\title{CCAMLR Marine Protected Areas}
\format{
Simple feature collection with 19 features and 13 fields:
\describe{
  \item{GAR_ID}{CCAMLR Statistical Areas}
  \item{Name}{CCAMLR Statistical Areas}
  \item{ShortLabel}{CCAMLR Statistical Areas}
  \item{LongLabel}{CCAMLR Statistical Areas}
  \item{StartDate}{CCAMLR Statistical Areas}
  \item{EndDate}{CCAMLR Statistical Areas}
}
}
\source{
CCAMLR \url{https://spatial.ccamlr.org/}
}
\usage{
ccamlr_mpa
}
\description{
CCAMLR Marine Protected Areas (MPAs).
}
\keyword{datasets}
