% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCSBT}
\alias{CCSBT}
\title{CCSBT Management Areas}
\format{
Simple feature collection with 12 features and 2 fields:
\describe{
  \item{id}{CCSBT Management Area}
  \item{Area}{CCSBT Management Area}
}
}
\source{
The CCSBT Data Manager \email{datamanager@ccsbt.org}
}
\usage{
CCSBT
}
\description{
The Commission for the Conservation of Southern Bluefin Tuna (CCSBT) Management Areas data set is a \code{sf} shapefile.
}
\keyword{datasets}
