% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cockle_QMA}
\alias{Cockle_QMA}
\title{New Zealand cockle QMAs}
\format{Simple feature collection with 13 features and 5 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
  \item{QmaName}{Quota Management Area or stock name}
}}
\source{
Fisheries New Zealand
}
\usage{
Cockle_QMA
}
\description{
New Zealand cockle Quota Management Areas (QMAs).
}
\keyword{datasets}
