% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ccamlr.R
\name{coord_ccamlr}
\alias{coord_ccamlr}
\title{CCAMLR coord}
\usage{
coord_ccamlr(proj = proj_ccamlr(), ...)
}
\arguments{
\item{proj}{The projection to use.}

\item{...}{Other arguments passed to \code{ggplot2::geom_sf()}.}
}
\value{
A CCAMLR \code{ggplot2} layer.
}
\description{
CCAMLR coord
}
\seealso{
\code{\link{geom_ccamlr}}
}
