% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depth_contour_polyline_hydro_1350k_11500k}
\alias{depth_contour_polyline_hydro_1350k_11500k}
\title{New Zealand depth contours 1350k by 11500k}
\format{Simple feature collection with 17157 features and 4 fields:
\describe{
  \item{depth}{the depth in m}
  \item{SCAMIN}{blah}
  \item{SORDAT}{blah}
  \item{SORIND}{blah}
}}
\source{
\url{https://data.linz.govt.nz/layer/50554-depth-contour-polyline-hydro-122k-190k/}
}
\usage{
depth_contour_polyline_hydro_1350k_11500k
}
\description{
New Zealand depth contours 1350k by 11500k.
}
\keyword{datasets}
