% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doc_marine_reserves}
\alias{doc_marine_reserves}
\title{New Zealand marine reserves}
\format{Simple feature collection with 44 features and 1 field:
\describe{
  \item{Name}{name of the marine reserve}
}}
\source{
\url{https://koordinates.com/layer/6026-doc-marine-reserves/}
}
\usage{
doc_marine_reserves
}
\description{
New Zealand Department of Conservation (DOC) marine reserves.
}
\keyword{datasets}
