% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exclusive_economic_zone_outer_limits_200_mile}
\alias{exclusive_economic_zone_outer_limits_200_mile}
\title{New Zealand Exclusive Economic Zone (EEZ)}
\format{Simple feature collection with 153 features and 29 fields:
\describe{
  \item{OBJECTID}{blah}
  \item{MARZONEID}{blah}
  \item{BP1}{blah}
}}
\usage{
exclusive_economic_zone_outer_limits_200_mile
}
\description{
New Zealand Exclusive Economic Zone (EEZ).
}
\keyword{datasets}
