% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FisheriesManagementAreas}
\alias{FisheriesManagementAreas}
\title{New Zealand Fisheries Management Areas (FMAs)}
\format{
Simple feature collection with 153 features and 29 fields:
\describe{
  \item{FeatureKey}{the}
  \item{LayerKey}{the}
  \item{LayerGroup}{the}
  \item{LayerName}{the}
  \item{MetadataKe}{the}
  \item{PortalUrl}{the}
  \item{SpeciesCode}{the}
}
}
\source{
Fisheries New Zealand
}
\usage{
FisheriesManagementAreas
}
\description{
New Zealand Fisheries Management Areas (FMAs).
}
\keyword{datasets}
