% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gebco_contours}
\alias{gebco_contours}
\title{GEBCO depth contours}
\format{
A simple feature collection with 1357601 features and 1 field.
}
\source{
https://www.opendem.info/download_bathymetry.html
}
\usage{
gebco_contours
}
\description{
NZ SST.
}
\keyword{datasets}
