% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gebco_depth_raster}
\alias{gebco_depth_raster}
\title{Gebco depth}
\format{Simple feature collection with 153 features and 29 fields:}
\source{
Gebco
}
\usage{
gebco_depth_raster
}
\description{
Gebco depth as a raster.
}
\keyword{datasets}
