% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gebco_NZ}
\alias{gebco_NZ}
\title{Gebco depth}
\format{
Simple feature collection with 153 features and 29 fields:
}
\source{
Gebco
}
\usage{
gebco_NZ
}
\description{
Gebco depth as a raster.
}
\keyword{datasets}
