% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ccamlr.R
\name{geom_ccamlr}
\alias{geom_ccamlr}
\title{CCAMLR geometries}
\usage{
geom_ccamlr(
  feature = "ssru",
  proj = proj_ccamlr(),
  fill = NA,
  colour = "black",
  ...
)
}
\arguments{
\item{feature}{A CCAMLR feature. Supported values are `ssru`, `label`,
`statistical_area`, `mpa`, `land`, and `gebco`.}

\item{proj}{The projection to use.}

\item{fill}{The fill colour for vector features.}

\item{colour}{The line or text colour for vector features.}

\item{...}{Other arguments passed to the selected ggplot2 geom.}
}
\value{
A ggplot2 layer containing the selected CCAMLR feature.
}
\description{
CCAMLR geometries
}
\seealso{
\code{\link{coord_ccamlr}}
}
