% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cra.R
\name{geom_cra}
\alias{geom_cra}
\title{Plot CRA labels}
\usage{
geom_cra(feature = "label", proj = proj_nzsf(), qma = NULL, area = NULL, ...)
}
\arguments{
\item{feature}{a spatial feature data.}

\item{proj}{The projection to use.}

\item{qma}{a spatial feature data.}

\item{area}{the field to rasterize.}

\item{...}{the field to rasterize.}
}
\value{
a \code{ggplot2} geom.
}
\description{
Plot CRA labels
}
