% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-coast.R
\name{get_coast}
\alias{get_coast}
\title{Get the New Zealand coastline}
\usage{
get_coast(proj = proj_nzsf(), resolution = "medium", keep = 1)
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{resolution}{the resolution, choose from "10", "50". "110", "150", "1250", "1500".}

\item{keep}{proportion of points to retain (0-1; default 1).}
}
\value{
New Zealand's coastline as an \code{sf} object.
}
\description{
Get the New Zealand coastline
}
\seealso{
\code{\link{plot_coast}}
}
