% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-brms.R
\name{get_conditional_smooths}
\alias{get_conditional_smooths}
\title{Plot a conditional_smooths output}
\usage{
get_conditional_smooths(
  x,
  resolution = 100,
  smooths = NULL,
  crs1 = st_crs(4326),
  crs2 = proj_nzsf(),
  ...
)
}
\arguments{
\item{x}{An object of class \code{brmsfit}.}

\item{resolution}{Number of support points used to generate 
the plots. Higher resolution leads to smoother plots. 
Defaults to \code{100}. If \code{surface} is \code{TRUE},
this implies \code{10000} support points for interaction terms,
so it might be necessary to reduce \code{resolution} 
when only few RAM is available.}

\item{smooths}{Optional character vector of smooth terms
to display. If \code{NULL} (the default) all smooth terms
are shown.}

\item{...}{Other arguments passed on to \code{conditional_smooths}.}
}
\value{
a \code{data.frame}.
}
\description{
Plot a conditional_smooths output
}
