% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-marine-reserves.R
\name{get_marine_reserves}
\alias{get_marine_reserves}
\title{Get Marine Reserves}
\usage{
get_marine_reserves(proj = proj_nzsf())
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}
}
\value{
New Zealand's marine reserves as an \code{sf} object.
}
\description{
Get Marine Reserves
}
\keyword{reserve}
