% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raster.R
\name{get_points_as_raster}
\alias{get_points_as_raster}
\title{Convert points to a raster}
\usage{
get_points_as_raster(data, field, fun = "sum", nrow = 100, ncol = 100)
}
\arguments{
\item{data}{a spatial feature data.}

\item{field}{the field to rasterize.}

\item{fun}{the function}

\item{nrow}{number of rows}

\item{ncol}{number of rows}
}
\value{
a raster.
}
\description{
Convert points to a raster
}
\examples{
x <- get_qma("CRA")
set.seed(4277)
pts <- sf::st_sf(sf::st_sample(x, size = 1000))
pts$z <- rnorm(nrow(pts))
r <- get_points_as_raster(data = pts, field = "z")

}
