% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fnz-standard-grid.R
\name{get_standard_grid}
\alias{get_standard_grid}
\title{Get Fisheries New Zealand standard grid definitions}
\usage{
get_standard_grid(
  cell_size,
  bounding_box,
  anchor = c(0, 0),
  return_raster = TRUE,
  crs = proj_nzsf(),
  square = TRUE
)
}
\arguments{
\item{cell_size}{square grid boundary length in km}

\item{bounding_box}{limits generated from call to \code{sf::st_bbox()}}

\item{anchor}{projected coordinates, in metres, used to anchor a shared grid
cell corner. The default is the EPSG:9191 origin at 175 degrees E,
40 degrees S.}

\item{return_raster}{return a raster or polygons}

\item{crs}{the CRS to use}

\item{square}{logical; if FALSE, create hexagonal grid}
}
\value{
New Zealand's standard grid polygon as a \code{sf} object or as a raster.
}
\description{
Get Fisheries New Zealand standard grid definitions
}
\seealso{
\code{\link{get_standard_grid_origin}}
}
\author{
Darcy Webber, Sophie Mormede, Charles Edwards
}
