% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fnz-standard-grid.R
\name{get_standard_grid_origin}
\alias{get_standard_grid_origin}
\title{Get Fisheries New Zealand standard grid origin}
\usage{
get_standard_grid_origin(cell_size, bounding_box, anchor = c(0, 0))
}
\arguments{
\item{cell_size}{square grid boundary length in km}

\item{bounding_box}{limits generated from call to \code{sf::st_bbox()}}

\item{anchor}{projected coordinates, in metres, used to anchor a shared grid
cell corner. The default is the EPSG:9191 origin at 175 degrees E,
40 degrees S.}
}
\value{
standard grid origin \code{data.frame}
}
\description{
Get Fisheries New Zealand standard grid origin
}
\seealso{
\code{\link{get_standard_grid}}
}
\author{
Darcy Webber, Sophie Mormede, Charles Edwards
}
