% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistical-areas.R
\name{get_statistical_areas}
\alias{get_statistical_areas}
\title{Get Statistical Areas}
\usage{
get_statistical_areas(area = "CRA", proj = proj_nzsf())
}
\arguments{
\item{area}{A fisheries area. Supported values include EEZ, CRA, FMA, JMA,
statistical areas, CCSBT, SIOFA, and SPRFMO.}

\item{proj}{The projection to use.}
}
\value{
New Zealand's statistical areas as an \code{sf} object.
}
\description{
Get Statistical Areas
}
\examples{
x <- get_statistical_areas(area = "CRA")
ggplot2::ggplot() +
  ggplot2::geom_sf(data = x, fill = NA)

}
\seealso{
\code{\link{plot_statistical_areas}}
}
