% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gisborne_TToR_Habitats}
\alias{Gisborne_TToR_Habitats}
\title{Gisborne Te Tapuwae o Rongokako Habitats}
\format{
Simple feature collection with 482 features and 15 fields:
\describe{
  \item{DEPTH}{the depth (m) of the habitat}
  \item{SUBSTRATE}{the type of subsrate}
  \item{Habitat}{the type of habitat}
  \item{TYPE}{the type of habitat}
}
}
\usage{
Gisborne_TToR_Habitats
}
\description{
Gisborne Te Tapuwae o Rongokako Habitats provided by the Department of Conservation (DOC).
}
\keyword{datasets}
