% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HAKE_QMA}
\alias{HAKE_QMA}
\title{New Zealand hake QMAs}
\format{Simple feature collection with 4 features and 4 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
}}
\source{
Fisheries New Zealand
}
\usage{
HAKE_QMA
}
\description{
New Zealand hake Quota Management Areas (QMAs).
}
\keyword{datasets}
