% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HOKI_QMA}
\alias{HOKI_QMA}
\title{New Zealand hoki QMAs}
\format{Simple feature collection with 2 features and 4 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
}}
\source{
Fisheries New Zealand
}
\usage{
HOKI_QMA
}
\description{
New Zealand hoki Quota Management Areas (QMAs).
}
\keyword{datasets}
