% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{international-fisheries-data}
\alias{international-fisheries-data}
\alias{SPRFMO}
\alias{SIOFA}
\alias{CCSBT}
\alias{ccamlr_eez}
\alias{ccamlr_mpa}
\alias{ccamlr_ssru}
\alias{ccamlr_statistical_areas}
\title{International fisheries areas}
\format{
Simple-feature objects: `SPRFMO` (8 features, 3 attributes),
  `SIOFA` (9, 29), `CCSBT` (12, 2), `ccamlr_eez` (7, 13), `ccamlr_mpa`
  (6, 12), `ccamlr_ssru` (63, 13), and `ccamlr_statistical_areas` (19, 6).

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 9 rows and 30 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 12 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 14 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 6 rows and 13 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 63 rows and 14 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 19 rows and 7 columns.
}
\source{
SPRFMO, SIOFA, CCSBT, and CCAMLR; CCAMLR spatial data are available
  from \url{https://spatial.ccamlr.org/}.
}
\usage{
SPRFMO

SIOFA

CCSBT

ccamlr_eez

ccamlr_mpa

ccamlr_ssru

ccamlr_statistical_areas
}
\description{
Snapshot statistical, management, and convention-area boundaries for
international fisheries organisations.
}
\keyword{datasets}
