% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JackMackerel_QMA}
\alias{JackMackerel_QMA}
\title{New Zealand jack mackerel QMAs}
\format{Simple feature collection with 4 features and 20 fields:
\describe{
  \item{SpeciesCod}{Species code}
  \item{SpeciesSci}{Scientific name}
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCom}{Common name}
}}
\source{
Fisheries New Zealand
}
\usage{
JackMackerel_QMA
}
\description{
New Zealand jack mackerel Quota Management Areas (QMAs).
}
\keyword{datasets}
