% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LING_QMA}
\alias{LING_QMA}
\title{New Zealand ling QMAs}
\format{Simple feature collection with 8 features and 4 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
}}
\source{
Fisheries New Zealand
}
\usage{
LING_QMA
}
\description{
New Zealand ling Quota Management Areas (QMAs).
}
\keyword{datasets}
