% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth.R
\name{lookup_depth}
\alias{lookup_depth}
\title{Look up GEBCO elevation at specific locations around New Zealand}
\usage{
lookup_depth(pts)
}
\arguments{
\item{pts}{An sf or sfc object containing point geometries.}
}
\value{
GEBCO elevation, in metres, at each location. Ocean depths are
  returned as negative elevations.
}
\description{
Look up GEBCO elevation at specific locations around New Zealand
}
\author{
Darcy Webber \email{darcy@quantifish.co.nz}
}
