% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mfe_average_sst}
\alias{mfe_average_sst}
\title{Gebco depth}
\format{
Simple feature collection with 153 features and 29 fields:

Simple feature collection with 153 features and 29 fields:
}
\source{
Gebco

Average SST

Average SST.

MFE
}
\usage{
mfe_average_sst
}
\description{
Gebco depth as a raster.
}
\keyword{datasets}
