% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz-fisheries-boundaries}
\alias{nz-fisheries-boundaries}
\alias{nz_general_statistical_areas}
\alias{nz_inshore_statistical_areas}
\alias{FisheriesManagementAreas}
\alias{exclusive_economic_zone_outer_limits_200_mile}
\alias{territorial_sea_outer_limit_12_mile}
\alias{rock_lobster_stat_areas}
\title{New Zealand fisheries boundaries and statistical areas}
\format{
Simple-feature objects with the following feature and attribute
  counts: `nz_general_statistical_areas` (121, 3),
  `nz_inshore_statistical_areas` (53, 3), `FisheriesManagementAreas` (25,
  11), `exclusive_economic_zone_outer_limits_200_mile` (153, 29),
  `territorial_sea_outer_limit_12_mile` (930, 4), and
  `rock_lobster_stat_areas` (43, 16).

An object of class \code{sf} (inherits from \code{data.frame}) with 121 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 53 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 12 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 153 rows and 30 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 930 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 43 rows and 17 columns.
}
\source{
Fisheries New Zealand fisheries maps
  \url{https://www.mpi.govt.nz/legal/legislation-standards-and-reviews/fisheries-legislation/maps-of-nz-fisheries}
  and Land Information New Zealand
  \url{https://data.linz.govt.nz/}
}
\usage{
nz_general_statistical_areas

nz_inshore_statistical_areas

FisheriesManagementAreas

exclusive_economic_zone_outer_limits_200_mile

territorial_sea_outer_limit_12_mile

rock_lobster_stat_areas
}
\description{
Snapshot boundary layers used by Fisheries New Zealand mapping helpers.
}
\keyword{datasets}
