% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz-habitat-data}
\alias{nz-habitat-data}
\alias{doc_marine_reserves}
\alias{Hauraki_Gulf_Marine_Park}
\alias{Gisborne_TToR_Habitats}
\alias{Gisborne_TToR_Reefs}
\alias{Rocky_reef_National_NZ}
\title{New Zealand protected areas and marine habitats}
\format{
Simple-feature objects: `doc_marine_reserves` has 50 features and
  one attribute; `Hauraki_Gulf_Marine_Park` has one feature and 19
  attributes; `Gisborne_TToR_Habitats` has 482 features and 15 attributes;
  `Gisborne_TToR_Reefs` has three features and one attribute; and
  `Rocky_reef_National_NZ` has 4,817 features and one attribute.

An object of class \code{sf} (inherits from \code{data.frame}) with 50 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 20 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 482 rows and 16 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4817 rows and 2 columns.
}
\source{
New Zealand Department of Conservation and source authorities
  recorded in `data-raw/README.md`.
}
\usage{
doc_marine_reserves

Hauraki_Gulf_Marine_Park

Gisborne_TToR_Habitats

Gisborne_TToR_Reefs

Rocky_reef_National_NZ
}
\description{
Snapshot protected-area and habitat layers used for contextual mapping.
}
\keyword{datasets}
