% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz-topographic-data}
\alias{nz-topographic-data}
\alias{nz_rivers}
\alias{nz_coastlines_topo_150k}
\alias{nz_coastlines_topo_1250k}
\alias{nz_coastlines_topo_1500k}
\alias{nz_coastlines_and_islands_polygons_topo_150k}
\alias{nz_coastlines_and_islands_polygons_topo_1250k}
\alias{nz_coastlines_and_islands_polygons_topo_1500k}
\alias{coastline_polyline_hydro_14k_122k}
\alias{depth_contour_polyline_hydro_122k_190k}
\alias{depth_contour_polyline_hydro_190k_1350k}
\alias{depth_contour_polyline_hydro_1350k_11500k}
\title{New Zealand coast, rivers, and depth contours}
\format{
Feature and attribute counts are: `nz_rivers` (4,866, 1),
  `nz_coastlines_topo_150k` (39, 2), `nz_coastlines_topo_1250k` (4, 2),
  `nz_coastlines_topo_1500k` (3, 1),
  `nz_coastlines_and_islands_polygons_topo_150k` (9,261, 7),
  `nz_coastlines_and_islands_polygons_topo_1250k` (1,699, 6),
  `nz_coastlines_and_islands_polygons_topo_1500k` (725, 2),
  `coastline_polyline_hydro_14k_122k` (5,043, 17),
  `depth_contour_polyline_hydro_122k_190k` (17,157, 4),
  `depth_contour_polyline_hydro_190k_1350k` (14,859, 4), and
  `depth_contour_polyline_hydro_1350k_11500k` (4,571, 4). Depth-contour
  objects include a `depth` attribute in metres.

An object of class \code{sf} (inherits from \code{data.frame}) with 4866 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 39 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 9261 rows and 8 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1699 rows and 7 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 725 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 5043 rows and 18 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 17157 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 14859 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4571 rows and 5 columns.
}
\source{
Land Information New Zealand
  \url{https://data.linz.govt.nz/}
}
\usage{
nz_rivers

nz_coastlines_topo_150k

nz_coastlines_topo_1250k

nz_coastlines_topo_1500k

nz_coastlines_and_islands_polygons_topo_150k

nz_coastlines_and_islands_polygons_topo_1250k

nz_coastlines_and_islands_polygons_topo_1500k

coastline_polyline_hydro_14k_122k

depth_contour_polyline_hydro_122k_190k

depth_contour_polyline_hydro_190k_1350k

depth_contour_polyline_hydro_1350k_11500k
}
\description{
Snapshot topographic and hydrographic simple-feature layers at several
source scales.
}
\keyword{datasets}
