% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_coastlines_topo_1250k}
\alias{nz_coastlines_topo_1250k}
\title{New Zealand coastlines 1250k}
\format{Simple feature collection with 9261 features and 7 fields:}
\source{
\url{https://data.linz.govt.nz/layer/51153-nz-coastlines-and-islands-polygons-topo-150k/}
}
\usage{
nz_coastlines_topo_1250k
}
\description{
New Zealand coastlines 1250k.
}
\keyword{datasets}
