% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_fisheries_general_statistical_areas}
\alias{nz_fisheries_general_statistical_areas}
\title{New Zealand General Statistical Areas}
\format{
Simple feature collection with 153 features and 29 fields:
}
\source{
Fisheries New Zealand

\url{https://data-mpi.opendata.arcgis.com/datasets/general-statistical-areas}
}
\usage{
nz_fisheries_general_statistical_areas
}
\description{
A more detailed description for the New Zealand general Statistical Areas can 
be found at \url{https://data-mpi.opendata.arcgis.com/datasets/general-statistical-areas}.
}
\details{

}
\keyword{datasets}
