% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_general_statistical_areas}
\alias{nz_general_statistical_areas}
\title{New Zealand General Statistical Areas}
\format{
Simple feature collection with 153 features and 3 fields:
\describe{
  \item{OBJECTID}{the object ID}
  \item{Statistica}{the statistical area}
  \item{Statisti_2}{the statistical area}
 }
}
\source{
Fisheries New Zealand \url{https://data-mpi.opendata.arcgis.com/datasets/general-statistical-areas}
}
\usage{
nz_general_statistical_areas
}
\description{
\code{nz_general_statistical_areas} returns a \code{sf} shapefile. A more detailed summary 
for the New Zealand General Statistical Areas can be found at 
\url{https://data-mpi.opendata.arcgis.com/datasets/general-statistical-areas}.
}
\keyword{datasets}
