% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_inshore_statistical_areas}
\alias{nz_inshore_statistical_areas}
\title{New Zealand Inshore Statistical Areas}
\format{
Simple feature collection with 53 features and 3 fields:
\describe{
  \item{OBJECTID}{the object ID}
  \item{Statistica}{the statistical area}
  \item{Statisti_2}{the statistical area}
 }
}
\source{
Fisheries New Zealand \url{https://data-mpi.opendata.arcgis.com/datasets/general-statistical-areas}
}
\usage{
nz_inshore_statistical_areas
}
\description{
A more detailed summary for the New Zealand inshore Statistical Areas can be 
found at \url{https://data-mpi.opendata.arcgis.com/datasets/general-statistical-areas}.
}
\keyword{datasets}
