% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Paua_QMA}
\alias{Paua_QMA}
\title{New Zealand paua QMAs}
\format{Simple feature collection with 11 features and 4 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
  \item{QmaName}{Quota Management Area or stock name}
}}
\source{
Fisheries New Zealand
}
\usage{
Paua_QMA
}
\description{
New Zealand paua Quota Management Areas (QMAs).
}
\keyword{datasets}
