% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pipi_QMA}
\alias{Pipi_QMA}
\title{New Zealand pipi QMAs}
\format{Simple feature collection with 10 features and 5 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
  \item{QmaName}{Quota Management Area or stock name}
}}
\source{
Fisheries New Zealand
}
\usage{
Pipi_QMA
}
\description{
New Zealand pipi Quota Management Areas (QMAs).
}
\keyword{datasets}
