% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coast.R
\name{plot_clip}
\alias{plot_clip}
\title{Clip to a shapefile.}
\usage{
plot_clip(x, ...)
}
\arguments{
\item{x}{The sf object to clip to.}

\item{...}{Other arguments passed on to \code{coord_sf}.}
}
\value{
a coord_sf.
}
\description{
Clip to a shapefile.
}
