% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot depth polylines around New Zealand}
\usage{
plot_depth(
  proj = proj_nzsf(),
  resolution = "low",
  depths = NULL,
  col_depth = FALSE,
  lty_depth = FALSE,
  ...
)
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{resolution}{the resolution.}

\item{depths}{a vector of specific depths to filter. Depths (in metres) that are available include: 0, 2, 5, 10, 20, 30, 50, 100, 200, 500, 1000, 2000, 3000, 4000, 5000, 6000, 7000, 8000, 9000, and 10000.}

\item{col_depth}{Different colours for the depth contours.}

\item{lty_depth}{Different line types for the depth contours.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
A \code{ggplot} object of New Zealand's depth polylines.
}
\description{
Plot depth polylines around New Zealand
}
\seealso{
\code{\link{get_depth}}
}
\author{
Darcy Webber \email{darcy@quantifish.co.nz}
}
