% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nz.R
\name{plot_nz}
\alias{plot_nz}
\title{Plot the New Zealand coastline.}
\usage{

  plot_nz(proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  resolution = "low", keep = 1, ...)
}
\arguments{
\item{proj}{The projection to use.}

\item{resolution}{the resolution}

\item{keep}{proportion of points to retain (0-1; default 0.05)}
}
\value{
ggplot of New Zealands coastline.
}
\description{
Plot the New Zealand coastline.
}
\examples{
ggplot() +
  plot_nz()

}
