% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qma.R
\name{plot_qma}
\alias{plot_qma}
\title{Plot a Quota Management Area (QMA)}
\usage{
plot_qma(qma = "CRA", proj = proj_nzsf(), ...)
}
\arguments{
\item{qma}{A Quota Management Area (QMA). Can be CRA, PHC, COC, ...}

\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
A \code{ggplot} object of the selected QMA.
}
\description{
This function is used to plot a Quota Management Area (QMA) as a \code{ggplot2} object.
}
\examples{
# Packhorse rock lobster
ggplot2::ggplot() +
  plot_qma(qma = "PHC")
  
# Jack mackerel
ggplot2::ggplot() +
  plot_qma(qma = "JMA")

# Paua
ggplot2::ggplot() +
  plot_qma(qma = "PAU")

}
\seealso{
\code{\link{get_qma}} to return Quota Management Area (QMA) polygons as \code{sf} objects.
}
