% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{proj_nzsf}
\alias{proj_nzsf}
\title{New Zealand projection}
\usage{
proj_nzsf()
}
\description{
The default projection for New Zealand. See the document "A standardised 
approach for creating spatial grids for New Zealand marine environment and 
species data". This CRS has been formally registered by EPSG as EPSG:9191 
(see: https://epsg.io/9191 which includes the formal OGP XML specification).
}
