% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qma-data}
\alias{qma-data}
\alias{Cockle_QMA}
\alias{HAKE_QMA}
\alias{HOKI_QMA}
\alias{JackMackerel_QMA}
\alias{LING_QMA}
\alias{OrangeRoughy_QMA}
\alias{OREO_QMA}
\alias{PackhorseRockLobster_QMA}
\alias{Paua_QMA}
\alias{Pipi_QMA}
\alias{Scallop_QMA}
\alias{Scampi_QMA}
\alias{SilverWarehou_QMA}
\alias{SouthernBlueWhiting_QMA}
\alias{SpinyRedRockLobster_QMA}
\alias{Squid_QMA}
\title{New Zealand Quota Management Areas}
\format{
Simple-feature objects with the following feature counts:
  `Cockle_QMA` (13), `HAKE_QMA` (4), `HOKI_QMA` (2),
  `JackMackerel_QMA` (4), `LING_QMA` (8), `OrangeRoughy_QMA` (8),
  `OREO_QMA` (5), `PackhorseRockLobster_QMA` (1), `Paua_QMA` (11),
  `Pipi_QMA` (10), `Scallop_QMA` (13), `Scampi_QMA` (11),
  `SilverWarehou_QMA` (4), `SouthernBlueWhiting_QMA` (5),
  `SpinyRedRockLobster_QMA` (10), and `Squid_QMA` (3). Objects have four
  attributes, except `Cockle_QMA`, `Pipi_QMA`, and `Scallop_QMA`, which have
  five.

An object of class \code{sf} (inherits from \code{data.frame}) with 13 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 2 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 11 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 13 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 11 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 3 rows and 5 columns.
}
\source{
Fisheries New Zealand fisheries maps
  \url{https://www.mpi.govt.nz/legal/legislation-standards-and-reviews/fisheries-legislation/maps-of-nz-fisheries}
}
\usage{
Cockle_QMA

HAKE_QMA

HOKI_QMA

JackMackerel_QMA

LING_QMA

OrangeRoughy_QMA

OREO_QMA

PackhorseRockLobster_QMA

Paua_QMA

Pipi_QMA

Scallop_QMA

Scampi_QMA

SilverWarehou_QMA

SouthernBlueWhiting_QMA

SpinyRedRockLobster_QMA

Squid_QMA
}
\description{
Snapshot Quota Management Area (QMA) boundaries. The common core fields are
`QMA`, `SpeciesCode`, `SpeciesScientific`, and the historically named
`SpeciesCommmon`; some shellfish layers also include `QmaName`.
}
\keyword{datasets}
