% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rock_lobster_stat_areas}
\alias{rock_lobster_stat_areas}
\title{New Zealand Rock Lobster Statistical Areas}
\format{Simple feature collection with 153 features and 29 fields:}
\source{
Fisheries New Zealand
}
\usage{
rock_lobster_stat_areas
}
\description{
New Zealand rock lobster Statistical Areas.
}
\keyword{datasets}
