% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rocky_reef_National_NZ}
\alias{Rocky_reef_National_NZ}
\title{New Zealand Rocky Reefs}
\format{Simple feature collection with 4817 features and 1 field:
\describe{
  \item{Source}{the source of the reef polygon}
}}
\usage{
Rocky_reef_National_NZ
}
\description{
New Zealand Rocky Reefs.
}
\keyword{datasets}
