% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Scampi_QMA}
\alias{Scampi_QMA}
\title{New Zealand scampi QMAs}
\format{
Simple feature collection with 1 features and 4 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
}
}
\source{
Fisheries New Zealand
}
\usage{
Scampi_QMA
}
\description{
New Zealand scampi Quota Management Areas (QMAs).
}
\keyword{datasets}
