% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Squid_QMA}
\alias{Squid_QMA}
\title{New Zealand squid QMAs}
\format{
Simple feature collection with 1 features and 4 fields:
\describe{
  \item{QMA}{Quota Management Area or stock}
  \item{SpeciesCode}{Species code}
  \item{SpeciesScientific}{Scientific name}
  \item{SpeciesCommon}{Common name}
}
}
\source{
Fisheries New Zealand
}
\usage{
Squid_QMA
}
\description{
New Zealand squid Quota Management Areas (QMAs).
}
\keyword{datasets}
