% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territorial_sea_outer_limit_12_mile}
\alias{territorial_sea_outer_limit_12_mile}
\title{New Zealand 12 mile limit}
\format{Simple feature collection with 153 features and 29 fields:}
\source{
Fisheries New Zealand
}
\usage{
territorial_sea_outer_limit_12_mile
}
\description{
New Zealand 12 mile limit.
}
\keyword{datasets}
