% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-nzsf.R
\name{theme_nzsf}
\alias{theme_nzsf}
\title{nzsf theme}
\usage{
theme_nzsf(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A ggplot2 theme for nzsf
}
