% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccamlr_eez}
\alias{ccamlr_eez}
\title{CCAMLR EEZ}
\format{
Simple feature collection with 7 features and 13 fields:
\describe{
  \item{LongLabel}{CCAMLR Statistical Areas}
}
}
\source{
CCAMLR \url{https://data.ccamlr.org/search/field_topic/reference-data-3}
}
\usage{
ccamlr_eez
}
\description{
CCAMLR EEZs.
}
\keyword{datasets}
