% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccamlr_ssru}
\alias{ccamlr_ssru}
\title{CCAMLR Small Scale Research Units}
\format{
Simple feature collection with 19 features and 13 fields:
\describe{
  \item{LongLabel}{CCAMLR Statistical Areas}
}
}
\source{
CCAMLR \url{https://data.ccamlr.org/search/field_topic/reference-data-3}
}
\usage{
ccamlr_ssru
}
\description{
CCAMLR SSRUs.
}
\keyword{datasets}
