% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCSBT}
\alias{CCSBT}
\title{CCSBT Management Areas}
\format{
Simple feature collection with 12 features and 2 fields:
\describe{
  \item{Area}{CCSBT Management Area}
}
}
\source{
CCSBT
}
\usage{
CCSBT
}
\description{
CCSBT Management Areas.
}
\keyword{datasets}
