% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ccamlr.R
\name{coord_ccamlr}
\alias{coord_ccamlr}
\title{CCAMLR coord}
\usage{
coord_ccamlr(proj = proj_ccamlr(), ...)
}
\arguments{
\item{proj}{The projection to use.}

\item{...}{A Quota Managemetn Area (QMA). Can be EEZ, CRA, JMA.}
}
\value{
New Zealands statistical areas as a \code{sf} object.
}
\description{
CCAMLR coord
}
\seealso{
\code{\link{geom_ccamlr}}
}
