% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ccsbt.R
\name{coord_ccsbt}
\alias{coord_ccsbt}
\title{CCSBT coord}
\usage{
coord_ccsbt(proj = proj_ccsbt(), ...)
}
\arguments{
\item{proj}{The projection to use.}

\item{...}{A Quota Managemetn Area (QMA). Can be EEZ, CRA, JMA.}
}
\value{
New Zealands statistical areas as a \code{sf} object.
}
\description{
CCSBT coord
}
\examples{
ggplot() +
  geom_ccsbt(feature = "area") +
  geom_ccsbt(feature = "land", fill = "black") +
  coord_ccsbt()

}
\seealso{
\code{\link{geom_ccsbt}}
}
