% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doc_marine_reserves}
\alias{doc_marine_reserves}
\title{New Zealand marine reserves}
\format{Simple feature collection with 44 features and 3 fields:
\describe{
  \item{Name}{blah}
  \item{Recorded_A}{blah blah}
  \item{Conservati}{blah blah}
}}
\source{
\url{https://koordinates.com/layer/6026-doc-marine-reserves/}
}
\usage{
doc_marine_reserves
}
\description{
New Zealand marine reserves.
}
\keyword{datasets}
