% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ccamlr.R
\name{geom_ccamlr}
\alias{geom_ccamlr}
\title{CCAMLR geometries}
\usage{
geom_ccamlr(
  feature = "ssru",
  proj = proj_ccamlr(),
  fill = NA,
  colour = "black",
  ...
)
}
\arguments{
\item{feature}{A Quota Managemetn Area (QMA). Can be EEZ, CRA, JMA.}

\item{proj}{The projection to use.}

\item{fill}{The projection to use.}

\item{colour}{The projection to use.}

\item{...}{The projection to use.}
}
\value{
New Zealands statistical areas as a \code{sf} object.
}
\description{
CCAMLR geometries
}
\examples{
ggplot() +
  geom_ccamlr(feature = "ssru") +
  geom_ccamlr(feature = "land", fill = "black") +
  coord_ccamlr()

}
\seealso{
\code{\link{coord_ccamlr}}
}
