% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ccsbt.R
\name{geom_ccsbt}
\alias{geom_ccsbt}
\title{CCSBT geometry}
\usage{
geom_ccsbt(
  feature = "area",
  proj = proj_ccsbt(),
  fill = NA,
  colour = "black",
  ...
)
}
\arguments{
\item{feature}{A CCSBT feature such as a management area, label or land.}

\item{proj}{The projection to use.}

\item{fill}{The fill colour for the selected feature.}

\item{colour}{The colour for the seleceted feature.}

\item{...}{Any additional arguments passed to \code{geom_sf}.}
}
\value{
A \code{geom_sf} object.
}
\description{
A helper function for obtaining a CCSBT related feature and returning a \code{geom_sf} object of that feature. Feature include the CCSBT 
management areas, labels or land. By default the projection returned by the function \code{proj_ccsbt} is used.
}
\examples{
ggplot() +
  geom_ccsbt(feature = "area") +
  geom_ccsbt(feature = "land", fill = "black") +
  coord_ccsbt()

}
\seealso{
\code{\link{coord_ccsbt}} for coord and \code{\link{proj_ccsbt}} for projection.
}
