% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ccsbt.R
\name{geom_ccsbt}
\alias{geom_ccsbt}
\title{CCSBT geometry}
\usage{
geom_ccsbt(
  feature = "area",
  proj = proj_ccsbt(),
  fill = NA,
  colour = "black",
  ...
)
}
\arguments{
\item{feature}{A Quota Managemetn Area (QMA). Can be EEZ, CRA, JMA.}

\item{proj}{The projection to use.}

\item{fill}{The projection to use.}

\item{colour}{The projection to use.}

\item{...}{The projection to use.}
}
\value{
New Zealands statistical areas as a \code{sf} object.
}
\description{
CCSBT geometry
}
\examples{
ggplot() +
  geom_ccsbt(feature = "area") +
  geom_ccsbt(feature = "land", fill = "black") +
  coord_ccsbt()

}
\seealso{
\code{\link{coord_ccsbt}}
}
