% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth.R
\name{geom_gebco}
\alias{geom_gebco}
\title{Plot GEBCO depth raster}
\usage{
geom_gebco(proj = proj_nzsf(), downsample = 3, ...)
}
\arguments{
\item{proj}{Projection.}

\item{downsample}{Downsampling rate: e.g. 3 keeps rows and cols 1, 4, 7, 10 etc.; a value of 0 does not downsample; can be specified for each dimension, e.g. c(5,5,0) to downsample the first two dimensions but not the third.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
A \code{ggplot} object of New Zealands depth polylines.
}
\description{
Plot GEBCO depth raster
}
\seealso{
\code{\link{get_depth}}
}
\author{
Darcy Webber \email{darcy@quantifish.co.nz}
}
