% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nz.R
\name{get_coast}
\alias{get_coast}
\title{Get the New Zealand coastline.}
\usage{
get_coast(keep = 0.05)
}
\arguments{
\item{keep}{proportion of points to retain (0-1; default 0.05)}
}
\value{
New Zealands coastline.
}
\description{
Get the New Zealand coastline.
}
\examples{
get_coast(keep = 0.25)

}
