% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coast.R
\name{get_coast}
\alias{get_coast}
\title{Get the New Zealand coastline.}
\usage{
get_coast(resolution = "low", keep = 1)
}
\arguments{
\item{resolution}{the resolution}

\item{keep}{proportion of points to retain (0-1; default 1)}
}
\value{
New Zealands coastline.
}
\description{
Get the New Zealand coastline.
}
\examples{
x <- get_coast(resolution = "low")
ggplot() +
  geom_sf(data = x, fill = "forestgreen")

}
