% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coast.R
\name{get_coast}
\alias{get_coast}
\title{Get the New Zealand coastline.}
\usage{

  get_coast(proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  resolution = "low", keep = 1)
}
\arguments{
\item{proj}{The projection to use.}

\item{resolution}{the resolution}

\item{keep}{proportion of points to retain (0-1; default 1)}
}
\value{
New Zealands coastline.
}
\description{
Get the New Zealand coastline.
}
\examples{
x <- get_coast(resolution = "low")
ggplot() +
  geom_sf(data = x, fill = "forestgreen")

}
