% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-depth.R
\name{get_depth}
\alias{get_depth}
\title{Get depth polylines around New Zealand}
\usage{
get_depth(proj = proj_nzsf(), resolution = "low", depths = NULL)
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{resolution}{the resolution.}

\item{depths}{a vector of specific depths to filter. Depths (in metres) that are available include: 0, 2, 5, 10, 20, 30, 50, 100, 200, 500, 1000, 2000, 3000, 4000, 5000, 6000, 7000, 8000, 9000, and 10000.}
}
\value{
New Zealands depth polylines as a \code{sf} object.
}
\description{
Get depth polylines around New Zealand
}
\examples{
x <- get_depth(resolution = "low")
ggplot() +
  geom_sf(data = x, colour = "lightblue")

}
\seealso{
\code{\link{plot_depth}}
}
\author{
Darcy Webber \email{darcy@quantifish.co.nz}
}
