% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-depth.R
\name{get_depth}
\alias{get_depth}
\title{Get depth polylines around New Zealand}
\usage{
get_depth(proj = proj_nzsf(), resolution = "low")
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{resolution}{the resolution.}
}
\value{
New Zealands depth polylines as a \code{sf} object.
}
\description{
Get depth polylines around New Zealand
}
\examples{
x <- get_depth(resolution = "low")
ggplot() +
  geom_sf(data = x, colour = "lightblue")

}
\seealso{
\code{\link{plot_depth}}
}
\author{
Darcy Webber \email{darcy@quantifish.co.nz}
}
