% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{Get depth polylines around New Zealand.}
\usage{
get_depth()
}
\value{
New Zealands depth polylines.
}
\description{
Get depth polylines around New Zealand.
}
\examples{
get_depth()

}
