% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{Get depth polylines around New Zealand}
\usage{

  get_depth(proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  resolution = "low")
}
\arguments{
\item{proj}{The projection to use.}

\item{resolution}{the resolution.}
}
\value{
New Zealands depth polylines.
}
\description{
Get depth polylines around New Zealand
}
\examples{
x <- get_depth(resolution = "low")
ggplot() +
  geom_sf(data = x, colour = "lightblue")

}
