% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{Get depth polylines around New Zealand.}
\usage{
get_depth(resolution = "low")
}
\arguments{
\item{resolution}{the resolution}
}
\value{
New Zealands depth polylines.
}
\description{
Get depth polylines around New Zealand.
}
\examples{
x <- get_depth(resolution = "low")
ggplot() +
  geom_sf(data = x, colour = "lightblue")

}
