% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-marine-reserves.R
\name{get_marine_reserves}
\alias{get_marine_reserves}
\title{Get Marine Reserves}
\usage{
get_marine_reserves(proj = proj_nzsf())
}
\arguments{
\item{proj}{The projection to use.}
}
\value{
New Zealands marine reserves.
}
\description{
Get Marine Reserves
}
\examples{
get_marine_reserves()

}
\keyword{reserve}
