% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marine_reserves.R
\name{get_marine_reserves}
\alias{get_marine_reserves}
\title{Get Marine Reserves}
\usage{

  get_marine_reserves(proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs")
}
\arguments{
\item{proj}{The projection to use.}
}
\value{
New Zealands marine reserves.
}
\description{
Get Marine Reserves
}
\examples{
get_marine_reserves()

}
\keyword{reserve}
