% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-qma.R
\name{get_qma}
\alias{get_qma}
\title{Get Quota Management Area (QMA) polygons}
\usage{
get_qma(qma = "CRA", proj = proj_nzsf())
}
\arguments{
\item{qma}{A Quota Managemetn Area (QMA).}

\item{proj}{The projection to use.}
}
\value{
A simple feature collection of QMA polygons.
}
\description{
Get Quota Management Area (QMA) polygons
}
\examples{
x <- get_qma(qma = "CRA")
ggplot() +
  geom_sf(data = x, fill = NA)

}
