% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qma.R
\name{get_qma}
\alias{get_qma}
\title{Get Quota Management Area (QMA) polygons}
\usage{
get_qma(qma = "CRA",
  proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs")
}
\arguments{
\item{qma}{A Quota Managemetn Area (QMA).}

\item{proj}{The projection to use.}
}
\value{
An \code{sf} \cod{data.frame} of QMA polygons.
}
\description{
Get Quota Management Area (QMA) polygons
}
\examples{
x <- get_qma(qma = "CRA")
ggplot() +
  geom_sf(data = x, fill = NA)

}
