% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fnz-standard-grid.R
\name{get_standard_grid}
\alias{get_standard_grid}
\title{Get Fisheries New Zealand standard grid definitions}
\usage{
get_standard_grid(x, bounding_box, clip_eez = TRUE)
}
\arguments{
\item{x}{square grid boundary length in km}

\item{bounding_box}{limits generated from call to sf::st_bbox()}

\item{clip_eez}{should the grid be clipped to the NZ EEZZ}
}
\value{
New Zealands standard grid polygon as a \code{sf} object.
}
\description{
Get Fisheries New Zealand standard grid definitions
}
\seealso{
\code{\link{get_standard_grid_origin}}
}
\author{
Sophie Mormede
}
