% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fnz-standard-grid.R
\name{get_standard_grid}
\alias{get_standard_grid}
\title{Get Fisheries New Zealand standard grid definitions}
\usage{
get_standard_grid(
  cell_size,
  bounding_box,
  anchor = c(0, 422600),
  return_raster = TRUE
)
}
\arguments{
\item{cell_size}{square grid boundary length in km}

\item{bounding_box}{limits generated from call to sf::st_bbox()}

\item{anchor}{the point to anchor the grid to}

\item{return_raster}{return a raster or polygons}
}
\value{
New Zealand's standard grid polygon as a \code{sf} object or as a raster.
}
\description{
Get Fisheries New Zealand standard grid definitions
}
\seealso{
\code{\link{get_standard_grid_origin}}
}
\author{
Darcy Webber, Sophie Mormede, Charles Edwards
}
