% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fnz-standard-grid.R
\name{get_standard_grid_origin}
\alias{get_standard_grid_origin}
\title{Get Fisheries New Zealand standard grid definitions}
\usage{
get_standard_grid_origin(x, bounding_box)
}
\arguments{
\item{x}{square grid boundary length in km}

\item{bounding_box}{limits generated from call to sf::st_bbox()}
}
\description{
Get Fisheries New Zealand standard grid definitions
}
\author{
Sophie Mormede
}
