% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-statistical-areas.R
\name{get_statistical_areas}
\alias{get_statistical_areas}
\title{Get Statistical Areas}
\usage{
get_statistical_areas(area = "CRA", proj = proj_nzsf())
}
\arguments{
\item{area}{A Quota Managemetn Area (QMA)}

\item{proj}{The projection to use.}
}
\value{
A FMA.
}
\description{
Get Statistical Areas
}
\examples{
x <- get_statistical_areas(area = "CRA")
ggplot() +
  geom_sf(data = x, fill = NA)

}
