% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistical_areas.R
\name{get_statistical_areas}
\alias{get_statistical_areas}
\title{Get Statistical Areas.}
\usage{
get_statistical_areas(area = "CRA")
}
\arguments{
\item{qma}{A Quota Managemetn Area (QMA)}
}
\value{
A QMA.
}
\description{
Get Statistical Areas.
}
\examples{
get_statistical_areas(area = "CRA")

}
