% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistical_areas.R
\name{get_statistical_areas}
\alias{get_statistical_areas}
\title{Get Statistical Areas.}
\usage{
get_statistical_areas(area = "CRA",
  proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs")
}
\arguments{
\item{proj}{The projection to use.}

\item{qma}{A Quota Managemetn Area (QMA)}
}
\value{
A FMA.
}
\description{
Get Statistical Areas.
}
\examples{
x <- get_statistical_areas(area = "CRA")
ggplot() +
  geom_sf(data = x, fill = NA)

}
