% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gisborne_TToR_Habitats}
\alias{Gisborne_TToR_Habitats}
\title{Gisborne Te Tapuwae o Rongokako Habitats}
\format{Simple feature collection with 482 features and 15 field:
\describe{
  \item{Habitat}{the type of habitat}
}}
\usage{
Gisborne_TToR_Habitats
}
\description{
Gisborne Te Tapuwae o Rongokako Habitats.
}
\keyword{datasets}
