% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-depth.R
\name{lookup_depth}
\alias{lookup_depth}
\title{Lookup the depth (m) at specific locations around New Zealand}
\usage{
lookup_depth(pts)
}
\arguments{
\item{pts}{The points to generate a depth for.}
}
\value{
Depth at each location.
}
\description{
Lookup the depth (m) at specific locations around New Zealand
}
\author{
Darcy Webber \email{darcy@quantifish.co.nz}
}
