% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mfe_average_sst}
\alias{mfe_average_sst}
\title{Average SST}
\format{Simple feature collection with 153 features and 29 fields:}
\source{
MFE
}
\usage{
mfe_average_sst
}
\description{
Average SST.
}
\keyword{datasets}
