% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_fisheries_general_statistical_areas}
\alias{nz_fisheries_general_statistical_areas}
\title{New Zealand General Statistical Areas}
\format{Simple feature collection with 153 features and 29 fields:}
\source{
Fisheries New Zealand
}
\usage{
nz_fisheries_general_statistical_areas
}
\description{
New Zealand general Statistical Areas.
}
\keyword{datasets}
