% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZBathymetry_2016_grid}
\alias{NZBathymetry_2016_grid}
\title{New Zealand depth ASCII grid}
\format{
A RasterLayer
}
\source{
NIWA \url{https://niwa.co.nz/our-science/oceans/bathymetry}
}
\usage{
NZBathymetry_2016_grid
}
\description{
New Zealand depth ASCII grid
}
\keyword{datasets}
