% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzsf.R
\docType{package}
\name{nzsf}
\alias{nzsf-package}
\alias{nzsf}
\title{New Zealand Spatial Features}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The nzsf package provides two main categories of important functions:
\code{get} and \code{plot}. Several other helper functions are also available, 
such as the \code{proj} functions which can be used to easily specify a useful projection to use.
}
\section{Get functions}{

The get functions retrieve bundled spatial data and perform requested
processing, such as transformation to a different projection. These
functions return simple-feature or raster objects.
}

\section{Plot functions}{

The plot functions combine a get function with an appropriate
\code{ggplot2} geom.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quantifish/nzsf}
  \item \url{https://www.quantifish.co.nz/nzsf/}
  \item Report bugs at \url{https://github.com/quantifish/nzsf/issues}
}

}
\author{
\strong{Maintainer}: Darcy Webber \email{darcy.webber@quantifish.co.nz}

Authors:
\itemize{
  \item Darcy Webber \email{darcy.webber@quantifish.co.nz}
}

}
\keyword{internal}
