% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzsf.R
\docType{package}
\name{nzsf}
\alias{nzsf}
\title{New Zealand Spatial Features}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The nzsf package provides two main categories of important functions:
\code{get} and \code{plot}. Several other helper functions are also available, 
such as the \code{proj} functions which can be used to easily specify a useful projection to use.
}
\section{Get functions}{

The get functions obtain shapefiles using a call to \code{data} and do some processing of that data (e.g. transform to a 
different projection). These functions return a spatial features object.
}

\section{Plot functions}{

The plot functions include a get function and an appropriate \code{ggplot2} geom to plot the shapefile.
}

