% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzsf.R
\docType{package}
\name{nzsf}
\alias{nzsf}
\alias{nzsf-package}
\title{nzsf: A package for computating the notorious bar statistic.}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The foo package provides three categories of important functions:
get, plot and baz.
}
\section{Get functions}{

The get functions ...
}

\section{Plot functions}{

The plot functions ...
}

