% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzsf.R
\docType{package}
\name{nzsf}
\alias{nzsf}
\alias{nzsf-package}
\title{New Zealand Spatial Features}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The nzsf package provides two categories of important functions:
\code{get} and \code{plot}.
}
\section{Get functions}{

The get functions obtain shapefiles using a call to \code{data} and do some processing of that data (e.g. transform to a 
different projection).
}

\section{Plot functions}{

The plot functions include a get function and an appropriate \code{ggplot2} geom to plot the shapefile.
}

