% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coast.R
\name{plot_clip}
\alias{plot_clip}
\title{Clip to a shapefile.}
\usage{
plot_clip(x, proj = proj_nzsf(), ...)
}
\arguments{
\item{x}{The sf object to clip to.}

\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{...}{Other arguments passed on to \code{coord_sf}.}
}
\value{
a coord_sf.
}
\description{
Clip to a shapefile.
}
