% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-coast.R
\name{plot_coast}
\alias{plot_coast}
\title{Plot the New Zealand coastline.}
\usage{
plot_coast(
  proj = proj_nzsf(),
  resolution = "medium",
  keep = 1,
  rivers = FALSE,
  ...
)
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{resolution}{the resolution, choose from "10", "50". "110", "150", "1250", "1500".}

\item{keep}{proportion of points to retain (0-1; default 1).}

\item{rivers}{Plot rivers over land as \code{geom_sf}.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
ggplot of New Zealands coastline.
}
\description{
Plot the New Zealand coastline.
}
\seealso{
\code{\link{get_coast}}
}
