% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coast.R
\name{plot_coast}
\alias{plot_coast}
\title{Plot the New Zealand coastline.}
\usage{
plot_coast(proj = proj_nzsf(), resolution = "low", keep = 1, ...)
}
\arguments{
\item{proj}{The projection to use.}

\item{resolution}{the resolution}

\item{keep}{proportion of points to retain (0-1; default 1)}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
ggplot of New Zealands coastline.
}
\description{
Plot the New Zealand coastline.
}
\examples{
ggplot() +
  plot_coast()

}
