% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-brms.R
\name{plot_conditional_smooths}
\alias{plot_conditional_smooths}
\title{Plot a conditional_smooths output}
\usage{
plot_conditional_smooths(x, resolution = 100, smooths = NULL, ...)
}
\arguments{
\item{...}{Other arguments passed on to \code{conditional_smooths}.}
}
\value{
a ggplot.
}
\description{
Plot a conditional_smooths output
}
