% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-depth.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot depth polylines around New Zealand}
\usage{
plot_depth(proj = proj_nzsf(), resolution = "low", ...)
}
\arguments{
\item{proj}{The projection to use.}

\item{resolution}{the resolution.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
ggplot of New Zealands depth polylines.
}
\description{
Plot depth polylines around New Zealand
}
\examples{
ggplot() + 
  plot_depth()

}
