% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot depth polylines around New Zealand.}
\usage{

  plot_depth(proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  ...)
}
\arguments{
\item{proj}{The projection to use.}
}
\value{
ggplot of New Zealands depth polylines.
}
\description{
Plot depth polylines around New Zealand.
}
\examples{
ggplot() + 
  plot_depth()

}
