% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-marine-reserves.R
\name{plot_marine_reserves}
\alias{plot_marine_reserves}
\title{Plot Marine Reserves}
\usage{
plot_marine_reserves(proj = proj_nzsf(), ...)
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
a \code{ggplot} object of New Zealand's marine reserves.
}
\description{
Plot Marine Reserves
}
\examples{
ggplot() +
  plot_marine_reserves()

}
