% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marine_reserves.R
\name{plot_marine_reserves}
\alias{plot_marine_reserves}
\title{Plot Marine Reserves.}
\usage{

  plot_marine_reserves(proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  ...)
}
\value{
ggplot of New Zealand's marine reserves.
}
\description{
Plot Marine Reserves.
}
\examples{
plot_marine_reserves()

}
