% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-qma.R
\name{plot_qma}
\alias{plot_qma}
\title{Get a Quota Managemetn Area.}
\usage{
plot_qma(qma = "CRA", proj = proj_nzsf(), ...)
}
\arguments{
\item{qma}{A Quota Managemetn Area (QMA). Can be CRA, PHC, COC, ...}

\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
A \code{ggplot} object of the selected QMA.
}
\description{
Get a Quota Managemetn Area.
}
\examples{
ggplot() + 
  plot_qma(qma = "CRA")

}
