% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qma.R
\name{plot_qma}
\alias{plot_qma}
\title{Get a Quota Managemetn Area.}
\usage{
plot_qma(qma = "CRA",
  proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  ...)
}
\arguments{
\item{qma}{A Quota Managemetn Area (QMA)}

\item{proj}{The projection to use.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
A ggplot of the selected QMA.
}
\description{
Get a Quota Managemetn Area.
}
\examples{
ggplot() + 
  plot_qma(qma = "CRA")

}
