% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-raster.R
\name{plot_raster}
\alias{plot_raster}
\title{Convert points to a raster}
\usage{
plot_raster(data, field, fun = "sum", nrow = 100, ncol = 100, ...)
}
\arguments{
\item{data}{a spatial feature data.}

\item{field}{the field to rasterize.}

\item{fun}{the function}

\item{nrow}{number of rows}

\item{ncol}{number of rows}

\item{...}{Other arguments passed on to \code{geom_raster}.}
}
\value{
a ggplot.
}
\description{
Convert points to a raster
}
\examples{
x <- get_qma("CRA")
pts <- st_sample(x, size = 1000) \%>\% 
  st_sf() \%>\% 
  mutate(z = rnorm(1:n()))
ggplot() +
  plot_raster(data = pts, field = "z")

}
