% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rivers.R
\name{plot_rivers}
\alias{plot_rivers}
\title{Plot New Zealand rivers.}
\usage{
plot_rivers(proj = proj_nzsf(), ...)
}
\arguments{
\item{proj}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
ggplot of New Zealands rivers.
}
\description{
Plot New Zealand rivers.
}
\examples{
ggplot() +
  plot_rivers()

}
