% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-statistical-areas.R
\name{plot_statistical_areas}
\alias{plot_statistical_areas}
\title{Get statistical areas}
\usage{

  plot_statistical_areas(proj = "+proj=aea +lat_1=-30 +lat_2=-50 +lat=-40 +lon_0=175 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  area = "CRA", ...)
}
\arguments{
\item{proj}{The projection to use.}

\item{area}{A Quota Managemetn Area (QMA)}

\item{...}{Other arguments passed on to \code{geom_sf}.}
}
\value{
A ggplot of the selected QMA.
}
\description{
Get statistical areas
}
\examples{
ggplot() + 
  plot_statistical_areas(area = "CRA")

}
