% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpinyRedRockLobster_QMA}
\alias{SpinyRedRockLobster_QMA}
\title{New Zealand red rock lobster QMAs}
\format{Simple feature collection with 6 features and 20 fields:
\describe{
  \item{FeatureKey}{blah}
  \item{LayerKey}{blah blah}
  \item{LayerGroup}{blah blah}
  \item{LayerName}{blah blah}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
SpinyRedRockLobster_QMA
}
\description{
New Zealand red rock lobster QMAs.
}
\keyword{datasets}
